# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A calorie tracking application backend with meal and workout tracking",
        "title": "BalanceLife API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/zhenyili/BalanceLife",
            "email": "support@balancelife.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/workouts/entries": {
            "get": {
                "description": "Returns workout entries for a user within a date range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workouts"
                ],
                "summary": "Get workout entries for a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.WorkoutEntry"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Logs a workout for a user with specified intensity and duration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workouts"
                ],
                "summary": "Create a new workout entry",
                "parameters": [
                    {
                        "description": "Workout entry details",
                        "name": "entry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.workoutEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.WorkoutEntry"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workouts/packages": {
            "get": {
                "description": "Returns a list of all workout packages, optionally filtered by goal type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workouts"
                ],
                "summary": "Get all workout packages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal type filter (LOSE, GAIN, ALL)",
                        "name": "goalType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.WorkoutPackage"
                            }
                        }
                    }
                }
            }
        },
        "/workouts/packages/{id}": {
            "get": {
                "description": "Returns details of a specific workout package",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workouts"
                ],
                "summary": "Get a workout package by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workout Package ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WorkoutPackage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.workoutEntryRequest": {
            "type": "object",
            "required": [
                "date",
                "durationMinutes",
                "intensityMultiplier",
                "packageId",
                "userId"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2023-03-18"
                },
                "durationMinutes": {
                    "type": "integer",
                    "maximum": 180,
                    "minimum": 5,
                    "example": 30
                },
                "intensityMultiplier": {
                    "type": "number",
                    "maximum": 2,
                    "minimum": 0.5,
                    "example": 1
                },
                "packageId": {
                    "type": "string",
                    "example": "workout1"
                },
                "userId": {
                    "type": "string",
                    "example": "usr1"
                }
            }
        },
...
        "models.GoalType": {
            "type": "string",
            "enum": [
                "LOSE",
                "GAIN",
                ""
            ],
            "x-enum-comments": {
                "GoalTypeAll": "Used for filtering all goal types"
            },
            "x-enum-varnames": [
                "GoalTypeLose",
                "GoalTypeGain",
                "GoalTypeAll"
            ]
        },
...
        "models.WorkoutEntry": {
            "type": "object",
            "properties": {
                "caloriesBurned": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "durationMinutes": {
                    "type": "integer"
                },
                "entryId": {
                    "type": "string"
                },
                "intensityMultiplier": {
                    "type": "number"
                },
                "packageId": {
                    "type": "string"
                },
                "timestamp": {
                    "description": "Used for querying by time range",
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.WorkoutPackage": {
            "type": "object",
            "properties": {
                "baseCaloriesBurn": {
                    "type": "integer"
                },
                "baseDurationMinutes": {
                    "type": "integer"
                },
                "caloriesBurnFormula": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "goalType": {
                    "description": "LOSE, GAIN, or BOTH",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.GoalType"
                        }
                    ]
                },
                "imageUrl": {
                    "type": "string"
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "packageId": {
                    "type": "string"
                },
                "workoutType": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}